final_pipeline/requirements.txt
feature_engine==1.8.2
icecream==2.1.3
isodate==0.7.2
matplotlib==3.10.0
numpy==1.24.1
pandas==2.2.3
Pillow==9.3.0
Pillow==11.0.0
Requests==2.32.3
scikit_learn==1.5.2
seaborn==0.13.2
sentence_transformers==3.3.1
shap==0.46.0
tqdm==4.67.1
ultralytics==8.3.49
